+++ src/codebase.php
@@ -7,10 +7,12 @@ $internal_trait_name_list = get_declared_traits();
 $internal_function_name_list = get_defined_functions()['internal'];
 
 if (file_exists(__DIR__ . '/../vendor/autoload.php')) {
-  // This is the normal path when Phan is installed only in the scope of a project.
+    // This is the path to autoload.php when Phan is installed globally.
+    // @phan-suppress-next-line PhanMissingRequireFile
     require_once __DIR__ . '/../vendor/autoload.php';
 } else {
-  // This is the path to autoload.php when Phan is installed globally.
+    // This is the normal path when Phan is installed only in the scope of a project.
+    // @phan-suppress-next-line PhanMissingRequireFile
     require_once __DIR__ . '/../../../autoload.php';
 }